+++ packages/notices/CHANGELOG.md
@@ -14,6 +14,10 @@
 
 - While `createNotice` only explicitly supported content of type `string`, it was not previously enforced. This has been corrected.
 
+### Bug Fixes
+
+- While `createNotice` only explicitly supported content of type `string`, it was not previously enforced. This has been corrected.
+
 ## 1.0.5 (2018-11-15)
 
 ## 1.0.4 (2018-11-09)